requirement.txt
appdirs==1.4.4
asgiref==3.2.10
Authlib==0.15.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.1.1
decorator==4.4.2
distlib==0.3.1
Django==3.1.2
django-filter==2.4.0
django-oauth2-login-client @ git+https://github.com/dragonfly-science/django-oauth2-login-client.git@4283819401c8f62ccbf3cdc1e34556ecc5d05470
django-rest-framework==0.1.0
djangorestframework==3.12.1
filelock==3.0.12
Flask==1.1.2
httpie==2.3.0
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.1
Markdown==3.3.1
MarkupSafe==1.1.1
networkx==2.5
nibabel==3.2.0
numpy==1.19.2
oauthlib==3.1.0
packaging==20.4
prov==1.5.3
psycopg2==2.8.6
pycparser==2.20
pydot==1.4.1
Pygments==2.7.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pyxnat==1.3
rdflib==5.0.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
scipy==1.5.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
virtualenv==20.0.31
Werkzeug==1.0.1
djangorestframework-simplejwt==4.6.0